refactor(api): migrate copilot components to common module
