--- src/qdash/api/lib/copilot_analysis.py
@@ -1,11 +1,0 @@
-"""Compatibility wrapper for shared copilot analysis models."""
-
-from qdash.common.copilot import analysis_models as _analysis
-
-AnalysisContextResult = _analysis.AnalysisContextResult
-AnalysisResponse = _analysis.AnalysisResponse
-AnalyzeRequest = _analysis.AnalyzeRequest
-BlocksResponse = _analysis.BlocksResponse
-ChatRequest = _analysis.ChatRequest
-ContentBlock = _analysis.ContentBlock
-TaskAnalysisContext = _analysis.TaskAnalysisContext
