+++ src/qdash/api/schemas/task_result.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from pydantic import BaseModel, ConfigDict, field_serializer, field_validator
 
-from qdash.api.lib.copilot_config import ModelConfig
+from qdash.common.copilot.settings import ModelConfig
 from qdash.common.datetime_utils import format_elapsed_time, parse_elapsed_time
 from qdash.datamodel.note import AiTriageReviewModel
 from qdash.datamodel.task import ParameterModel
