+++ tests/qdash/api/lib/test_copilot_agent.py
@@ -5,8 +5,8 @@ from typing import Any
 
 import pytest
 
-from qdash.api.lib.copilot_agent import _parse_response, _run_chat_completions
-from qdash.api.lib.copilot_config import CopilotConfig, ModelConfig
+from qdash.common.copilot.llm_agent import _parse_response, _run_chat_completions
+from qdash.common.copilot.settings import CopilotConfig, ModelConfig
 
 
 def test_parse_response_accepts_review_triage_markdown_without_json() -> None:
