+++ tests/qdash/api/services/test_issue_knowledge_service.py
@@ -845,7 +845,7 @@ class TestCallLlmExtract:
         config = MagicMock()
         config.enabled = False
 
-        with patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=config):
+        with patch("qdash.common.copilot.settings.load_copilot_config", return_value=config):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread text"
             )
@@ -870,8 +870,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "T1 issue", "Thread"
@@ -891,8 +891,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
@@ -910,8 +910,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
@@ -926,8 +926,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(side_effect=RuntimeError("API unreachable"))
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
@@ -944,8 +944,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
@@ -965,8 +965,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
@@ -983,8 +983,8 @@ class TestCallLlmExtract:
         mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
 
         with (
-            patch("qdash.api.lib.copilot_config.load_copilot_config", return_value=mock_config),
-            patch("qdash.api.lib.copilot_agent._build_client", return_value=mock_client),
+            patch("qdash.common.copilot.settings.load_copilot_config", return_value=mock_config),
+            patch("qdash.common.copilot.llm_agent._build_client", return_value=mock_client),
         ):
             result = await IssueKnowledgeService._call_llm_extract(
                 "CheckT1", "chip-A", "Q0", "Title", "Thread"
